grc: Add optional parameter to pad_* blocks description
Pad sink/sources blocks may be optional, so connections in the calling
flowgraph may not be required.
But at the moment setting optional to yes in the pad property window has
no effect. The pad block is alwas required.
This is due to an error in the yml description of pad* blocks.