grc: update disabled blocks if they depend on others
If we want disabled blocks to be updated properly,
we have to consider all blocks when updating the fg not only the enabled.

This does the patch of Flowgraph.py and fixes #4788.

But a problem remains.
If you use the example in #4788 and disable the symbol_rate block,
only the samp_rate block 'turns to red'.
The other blocks don't change.
Only if you open and close an arbitrary block, the other blocks are updated.

This is fixed by clearing the namespace in renew_namespace in FlowGraph.py.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>

grc: update disabled blocks

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>