grc: Fix category and module tooltips
Category and module tooltips have been broken since GNU Radio 3.8.
The _format_cat_tooltip function expects a tuple representing the
module & category hierarchy, but a single string is passed in. As
a result, everyhing is treated as a category, and only the last
letter of the category or module name is displayed. I've fixed
the problem by passing in the full tuple instead.

Signed-off-by: Clayton Smith <argilo@gmail.com>