grc: Add aliases for "sc16==short" and "sc8==byte".
When type checking was added for GRC connections, the valid connections
between a block with the "sc16" type (e.g. UHD blocks) and a block
with the "short" type and a vector length of 2 (e.g. ishort->complex)
now fail. Adding these aliases fixes that as a stop-gap until a better
solution can be found.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>