grc: Look up type aliases as a set instead of a single value.
After #5127, we inadvertently had duplicate dictionary keys, meaning the
last entry is the only one that actually existed and some desired
aliases were missing. This changes the ALIAS_OF dictionary to ALIASES_OF
where the values are now a set of aliases instead of a single string
value. The one use of the ALIAS_OF dictionary was changed to operate on
the returned set.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>